Always give a timeout to thread join, to avoid uninterruptible sleep
In Python 2.x, a thread join with no timeout is uninterruptible.
This means a process could get stuck in executor shutdown such that
the process can't be killed by SIGTERM.  That's undesirable;
provide a timeout so that it's possible to interrupt.

Python 3 doesn't have this problem (tested on 3.6) - thread joins
are interruptible there.

This probably doesn't entirely fix the problem for Python 2, for
example ThreadPoolExecutor can probably still hang uninterruptably
at shutdown, but at least my code won't be contributing to the
problem...

Fixes #42