Revise TimeoutExecutor to "cancel on timeout"
The original concept of TimeoutExecutor was simply to apply a
default timeout on calls to result/exception.  The new concept
is instead to cancel futures which don't complete within the
timeout.

I think this makes more sense as the previous concept couldn't
work with methods such as `wait`, `as_completed`, which don't
work by calling result/exception.

The docs have been updated to pretend that TimeoutExecutor was
introduced in 1.7.0, since the semantics changed so drastically.

Closes #50