Remove one test on "cancel" behavior
This test was checking that, if a future could not be cancelled,
it was either running or done.

Although true for the executor/future implementations in
concurrent.futures and probably intended to be true in general,
this is impractical to achieve for some of the executors in this
module.  Previously, certain executors were whitelisted in the
test as being allowed to bend the rule.

I want to add more which also can't satisfy this condition, which
would require updating the test, but it seems to me that the test
is not worthwhile; it rather makes more sense to test all the
executors consistently and accept that this condition is never
guaranteed.