CI: fix testing for externally submitted PRs
I had expected that environment variables in a given `context`
are simply omitted for runs triggered by an unauthorized user,
but instead the build simply fails.  To fix this, move the
configurations which need a context into post-merge trigger only.

The downside is that docs building is no longer tested prior to merge.