Permit various arguments-differ pylint warnings
pylint rightfully complains that the signature of Executor.submit
is not matched in various subclasses. This is a warning which has
recently appeared with no related changes in the code itself.

I've tried to fix it, but the error seems to persist and may differ
on different python versions (did concurrent.futures.Executor
change perhaps?)

For now, let's mark it as permitted and file an issue to investigate
properly later.