CI: add reverse dependency tests
To give more confidence that changes don't introduce regressions
or incompatibilities, let's run the tests of some *other* projects
which depend on us.

This is a bit of an experiment and I'm not sure how well it will
work in practice.  It depends on the other projects keeping their
tests entirely stable, and it might run through the CircleCI
quota pretty quick.