Improve performance/limits of various functions
These functions operate on collections of futures:

- f_sequence
- f_zip
- f_and
- f_or

The previous implementations of these relied on recursion. This
meant the functions would consume O(N) stack frames where N is
the number of futures provided as input. In practice, this caused
all of the functions to stop working at around ~1,000 inputs.

That's not enough for some important use-cases, so let's rewrite
them to avoid recursion and raise the limits considerably. The new
implementations are verified as working with 100,000 inputs, and
this limit has been documented.