map: don't apply any default error_fn
Letting error_fn default to a function which re-raises the exception
simplified the code, but unfortunately those re-raises end up in the
backtraces seen by user code, and will add one stack frame per
map. This is ugly and possibly misleads users into thinking something
is going wrong, so let's not do it that way.

Fixes #169