pylint: disable signature-differs
In Python 3.9, a new keyword argument was added to shutdown().
We will run into this warning everywhere due to that.
It's not yet clear what to do about it, since we can't easily
match the signature both prior to and after that python version.
Let's disable this for now to unblock CI.