Introduce dependabot & pin dependencies
Let's use dependabot to ensure the CI for this project is regularly
exercised and any problems in dependencies are detected.

It is expected that at least one Python3 test configuration uses
dependencies kept as up-to-date as possible by dependabot.
The Python2 configurations however don't use dependency pinning
as there is no clear way to have dependabot running pip-compile for
both Python2 and Python3 when making any changes.