Reduce reference cycles to improve memory usage
Previously, a MapFuture and its delegate would always form a reference
cycle. If the futures point to large objects, this can result in
significant memory pressure as the cycle will not be freed until GC gets
around to it. Since GC is only triggered according to object counts and
not object sizes, this can increase apparent memory usage a lot in some
cases.

The cycle can be broken as soon as the delegate future has resolved, so
let's do that. (RetryFuture and PollFuture were already doing this, so
it seems there was an oversight specifically for MapFuture.)

Fixes #344